refactoring changes induced by previous commit
changes not reviewed yet
