+++ docs/node tutorials/doc_examples/custom_widgets/widgets.py
@@ -1,4 +1,4 @@
-from ryven.NWENV import *
+from ryven.gui_env import *
 
 from qtpy.QtWidgets import QWidget, QVBoxLayout, QLabel, QPushButton, QLineEdit
