+++ docs/node tutorials/doc_examples/mypackage/nodes.py
@@ -1,4 +1,4 @@
-from ryven.NENV import *
+from ryven.node_env import *
 
 
 class NodeBase(Node):
