+++ ryven/example_nodes/OpenCV/widgets.py
@@ -1,4 +1,4 @@
-from ryven.NWENV import *
+from ryven.gui_env import *
 from qtpy.QtWidgets import QLabel, QPushButton, QFileDialog, QVBoxLayout, QWidget, QTextEdit
 from qtpy.QtGui import QImage, QPixmap, QFont
 from qtpy.QtCore import Signal, QSize, QTimer
