+++ ryven/example_nodes/std/control_structures.py
@@ -1,4 +1,4 @@
-from ryven.NENV import *
+from ryven.node_env import *
 
 
 class CSNodeBase(Node):
