+++ ryven/example_nodes/std/special_nodes.py
@@ -1,8 +1,8 @@
 import code
 from contextlib import redirect_stdout, redirect_stderr
 
-from ryven.NENV import *
-widgets = import_widgets(__file__)
+from ryven.node_env import *
+widgets = import_guis(__file__)
 
 
 class NodeBase(Node):
